renamed from [...slug] to [slug] - the rest is formatting noise! also added a link to tsconfig to stop vscode from complaining about the template tag
